Refactor: Remove default values from UiState classes
This commit removes default values from the following UiState classes:
- `EditProfileUiState`
- `StatsUiState`
- `CreateQuestUiState`
- `FirstSetupUiState`
- `HabitsUiState`
- `CreateHabitUiState`
- `ViewProfileUiState`
- `MainUiState`

Default values are now set in their respective ViewModels.
In `ViewProfileViewModel`, the `displayName` is now initialized with the app name string resource.
In `MainViewModel`, the initial `MainUiState` is set with default values and updated when `appUser` data is collected.
In `FirstSetupViewModel`, `imageUri` is now correctly parsed using `toUri()`.