feat(create-quest): Pre-select category when creating a quest
This commit updates the `CreateQuestScreen` to allow for pre-selecting a quest category upon navigation. This is achieved by passing an optional `selectedCategoryIndex` to the `CreateQuestRoute`.

The `CreateQuestViewModel` has been refactored to use a Hilt `@AssistedFactory` to receive this index. If a `selectedCategoryIndex` is provided, the corresponding category will be automatically selected when the screen loads.

- **`CreateQuestViewModel.kt`**:
    - Implemented an `@AssistedFactory` to inject the optional `selectedCategoryIndex`.
    - The `SavedStateHandle` is no longer used to retrieve this parameter.

- **`CreateQuestScreen.kt`**:
    - Updated the `CreateQuestViewModel` instantiation to use the new assisted factory, passing the `selectedCategoryIndex` from the route.

- **`ActivityMain.kt`**:
    - Modified the `CreateQuestRoute` navigation entry to pass the `selectedCategoryIndex` to `CreateQuestScreen`.